Renamed AddNewAlgorithmScript to AddNewAlgoritm.
